requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-admin==2.0.1
django-cors-headers==3.13.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-postgresql==0.0.3
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.13.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
pipreqs==0.4.11
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.6
python-dotenv==0.20.0
Python-Rest-Framework==0.3.14
pytz==2022.1
requests==2.28.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
xlwt==1.3.0
yarg==0.1.9
